requirements.txt
flask
numpy
opencv-python
pillow
pyzbar